LayoutTests/imported/w3c/web-platform-tests/navigation-timing/nav2-test-timing-persistent-expected.txt



PASS iframe navigation times are persistent after the iframe is removed. Part 1.
FAIL iframe navigation times are persistent after the iframe is removed Part 2. assert_equals: Only one navigation time entry expected 1 but got 0
